src/test/java/wniemiec/mobilex/ama/parser/screens/behavior/instruction/BreakInstructionJsonParserTest.java
package wniemiec.mobilex.ama.parser.screens.behavior.instruction;

import java.io.IOException;
import org.json.JSONObject;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import wniemiec.mobilex.ama.models.behavior.Instruction;
import wniemiec.mobilex.ama.parser.exception.ParseException;


class BreakInstructionJsonParserTest {
    
    //-------------------------------------------------------------------------
    //		Attributes
    //-------------------------------------------------------------------------
    private BreakInstructionJsonParser parser;
    private Instruction parsedInstruction;
    private JSONObject label;


    //-------------------------------------------------------------------------
    //		Test hooks
    //-------------------------------------------------------------------------
    @BeforeEach
    void setUp() {
        parser = BreakInstructionJsonParser.getInstance();
        parsedInstruction = null;
        label = null;
    }


    //-------------------------------------------------------------------------
    //		Tests
    //-------------------------------------------------------------------------
    @Test
    void testParseWithLabel() throws ParseException, IOException {
        withLabel(buildIdentifier("foo"));
        doParsing();
        assertParsedCodeIs("break foo");
    }


    //-------------------------------------------------------------------------
    //		Methods
    //-------------------------------------------------------------------------
    private JSONObject buildIdentifier(String name) {
        JSONObject expression = new JSONObject();

        expression.put("type", "Identifier");
        expression.put("name", name);

        return expression;
    }

    private void withLabel(JSONObject label) {
        this.label = label;
    }

    private void doParsing() throws ParseException, IOException {
        parsedInstruction = parser.parse(buildContinueStatement());
    }

    private JSONObject buildContinueStatement() {
        JSONObject instruction = new JSONObject();

        instruction.put("type", "BreakStatement");
        instruction.put("label", label);

        return instruction;
    }

    private void assertParsedCodeIs(String code) {
        assertHasSameLine(code, parsedInstruction.toCode());
    }

    private void assertHasSameLine(String expected, String obtained) {
        Assertions.assertEquals(
            removeWhiteSpaces(expected),
            removeWhiteSpaces(obtained)
        );
    }

    private String removeWhiteSpaces(String text) {
        return text.replaceAll("[\\s\\t]+", "");
    }
}
